refactor(install): P1.L1 -- extract validation.py from install.py
Moves three leaf validation helpers (package-existence checks, local-path
diagnostics) into the install engine package as
`apm_cli/install/validation.py`.  These functions had zero coupling to
the rest of install.py -- they take packages and return booleans or
strings -- so this is a pure relocation.

Functions moved:
- _validate_package_exists
- _local_path_failure_reason
- _local_path_no_markers_hint

`_validate_and_add_packages_to_apm_yml` remains in commands/install.py
because it calls _validate_package_exists and _local_path_failure_reason
via module-level name lookup, and 30+ tests patch
`apm_cli.commands.install._validate_package_exists` to intercept those
calls.  Keeping the orchestrator co-located with the re-exported names
preserves all @patch targets without any test modifications.

`commands/install.py` re-exports the three moved names so existing test
patches keep working verbatim.  Targeted install tests (175) + full unit
suite (3972) green.

install.py LOC: 2905 -> 2630.
